requirements.txt
# Core Libraries
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
tqdm==4.66.2

# Machine Learning & NLP
sentence-transformers==2.5.1
spacy==3.6.1
transformers==4.38.2
torch==2.1.0
accelerate

# Streamlit for UI
streamlit==1.33.0

# Google Gemini AI API
google-generativeai==0.3.1

# Additional Utilities
bitsandbytes
jupyter
wheel
